requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
redis==5.0.1
pyyaml==6.0.1
prometheus-client==0.19.0
pydantic==2.5.0
psutil==5.9.6